requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --resolver=backtracking requirements/requirements.in
#
asgiref==3.6.0
    # via django
attrs==22.2.0
    # via pytest
cfgv==3.3.1
    # via pre-commit
click==8.1.3
    # via django-click
coverage[toml]==7.2.3
    # via pytest-cov
distlib==0.3.6
    # via virtualenv
dj-database-url==1.3.0
    # via environs
dj-email-url==1.0.6
    # via environs
django==4.2
    # via
    #   -r requirements/requirements.in
    #   dj-database-url
    #   model-bakery
django-cache-url==3.4.4
    # via environs
django-click==2.3.0
    # via -r requirements/requirements.in
django-test-plus==2.2.1
    # via -r requirements/requirements.in
environs[django]==9.5.0
    # via -r requirements/requirements.in
exceptiongroup==1.1.1
    # via pytest
filelock==3.11.0
    # via virtualenv
identify==2.5.22
    # via pre-commit
iniconfig==2.0.0
django-hashid-field==3.3.7
    # via pytest
marshmallow==3.19.0
mysqlclient==2.1.1
    # via environs
model-bakery==1.11.0
    # via -r requirements/requirements.in
nodeenv==1.7.0
    # via pre-commit
packaging==23.0
    # via
    #   marshmallow
    #   pytest
platformdirs==3.2.0
    # via virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==3.2.2
    # via -r requirements/requirements.in
psycopg2-binary==2.9.6
    # via -r requirements/requirements.in
pytest==7.2.2
    # via
    #   -r requirements/requirements.in
    #   pytest-cov
    #   pytest-django
pytest-cov==4.0.0
    # via -r requirements/requirements.in
pytest-django==4.5.2
    # via -r requirements/requirements.in
python-dotenv==1.0.0
    # via environs
pyyaml==6.0
    # via pre-commit
sqlparse==0.4.3
    # via django
tomli==2.0.1
    # via
    #   coverage
    #   pytest
typing-extensions==4.5.0
    # via dj-database-url
virtualenv==20.21.0
    # via pre-commit
whitenoise==6.4.0
    # via -r requirements/requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
redis==4.6.0
hiredis==2.2.3
